// frontend-app/src/apiClient/models/ForestProjectTokenHolder.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { SecurityTokenContractType } from "./SecurityTokenContractType";

export type ForestProjectTokenHolder = {
	cis2_address: string;
	token_id: string;
	holder_address: string;
	frozen_balance: string;
	un_frozen_balance: string;
	forest_project_id: string;
	forest_project_name: string;
	contract_type: SecurityTokenContractType;
	email: string;
};